Create Time for Action
To drive change. To urgently build and improve justice systems that ensure fairness, promote safety, and strengthen communities.

Achieving Our Mission
We work with others who share our vision to tackle the most pressing injustices of our day—from the causes and consequences of mass incarceration, racial disparities, and the loss of public trust in law enforcement, to the unmet needs of the vulnerable, the marginalized, and those harmed by crime and violence.

LEARN ABOUT OUR PRIORITIES
In its scale and brutality, the American justice system is a global aberration.
The United States has
5%
of world's population
25%
of world's incarcerated
The United States incarcerates more people than any other country in the world.
People of color make up
30%
of U.S. population
60%
of U.S. prison population
People of color are incarcerated at an unmistakably higher rate than white people.
How We Work
Vera works in partnership with local, state, and national government officials to create change from within.

We study problems
that impede human dignity and justice.
We pilot solutions
that are at once transformative and achievable.
We engage diverse communities
in informed debate.
We harness the power of evidence
to drive effective policy and practice.
Our Values
Respect
We listen closely, accept and offer feedback constructively, invite diverse perspectives, and treat each other and our partners with respect and compassion.

Independence
We ask difficult questions, entertaining unconventional answers, and reckoning with any uncomfortable truths that our research and practice may reveal.

Collaboration
We actively engage others—from those denied justice to those charged with its administration—because we are most effective when we work in partnership.

Commitment
We dedicate ourselves to the highest standards in research and practice because nothing short of continual learning and improvement will help us effect the change we seek.

Race Equity
We commit to the ongoing and necessary work of disrupting and eliminating structural racism and white supremacy, internally and externally, in order to fully realize race equity in our justice system and beyond.

Where We Work
Working with government and civil leaders to improve justice systems, Vera's projects are active in more than 40 states.

Our History
In 1961, philanthropist Louis Schweitzer and magazine editor Herb Sturz recognized the injustice of a bail system in New York City that locked people up simply for being poor. Working with criminal justice leaders, they explored the problem from many angles, setting their sights on developing a practical, innovative solution to NYC’s overreliance on cash bail. Within a few years, they had demonstrated that New Yorkers too poor to afford bail but with strong ties to their communities could be released and still show up for trial. The experiment—dubbed the Manhattan Bail Project—resulted in dozens of similar projects in cities around the country, and in landmark legislative reform of the federal bail system. It also led to the founding of the Vera Institute of Justice—named for Schweitzer's inspiring mother—to pursue similar initiatives. Today, Vera is working on more than five dozen projects that aim to unravel the impediments to human dignity and justice, while changing the lives of individuals. More than half a century after the Manhattan Bail Project, Vera embraces new tools—like the power of mining big data to unearth injustice, the potential of competitions to seek out the most motivated leaders, and the importance of communications to engage with our ideas. These are all part of an approach that has remained remarkably consistent, characterized by a disciplined focus on specific problems, partnering with public sector leaders willing to work with us to seek change, and a reliance on rigorous research every step of the way. Underlying this unique approach is an abiding optimism that even the most troubled systems can transform. The result: justice systems that ensure fairness, promote safety, and strengthen communities..